requirements.txt
requests
kafka-python
pyspark>=3.3.0
folium
streamlit
joblib
psycopg2-binary
pandas
scikit-learn
streamlit-folium
pandas
requests
pytest